integrated flashdrive setting, debugging collection

+++ src/py/poll/location.py
@@ -5,7 +5,7 @@ import os
 def manage_output_location():
     currently_outputting_to_flashdrive = False
     while True:
-        if os.path.exists("/media/pi/SUNDAR_RESEARCH"):
+        if get_flashdrive_setting and os.path.exists("/media/pi/SUNDAR_RESEARCH"):
             if not currently_outputting_to_flashdrive:
                 with open("flags/destination", "w") as output_location_flag:
                     output_location_flag.write("/media/pi/SUNDAR_RESEARCH/data/")
@@ -16,6 +16,13 @@ def manage_output_location():
                     output_location_flag.write("data/")
                 currently_outputting_to_flashdrive = False
         time.sleep(1)
+
+def get_flashdrive_setting():
+    with open("settings/flashdrive", "r") as flashdrive_file:
+        value = flashdrive_file.read()
+    if value == "true":
+        return True
+    return False
 
 # retrieves the string in the destination flag file
 def get_output_location():
